+++ src/main/kotlin/no/nav/familie/pdf/pdf/pdfElementer/SpørsmålOgSvar.kt
@@ -3,9 +3,9 @@ package no.nav.familie.pdf.pdf.pdfElementer
 import com.itextpdf.kernel.pdf.tagging.StandardRoles
 import com.itextpdf.layout.element.Paragraph
 import com.itextpdf.layout.element.Text
-import no.nav.familie.pdf.pdf.PdfUtils.FontStil
-import no.nav.familie.pdf.pdf.PdfUtils.settFont
+import no.nav.familie.pdf.pdf.FontStil
 import no.nav.familie.pdf.pdf.domain.VerdilisteElement
+import no.nav.familie.pdf.pdf.settFont
 
 fun lagSpørsmålOgSvar(element: VerdilisteElement): Paragraph =
     Paragraph().apply {
