+++ src/main/kotlin/no/nav/familie/pdf/pdf/pdfElementer/Tekst.kt
@@ -3,8 +3,8 @@ package no.nav.familie.pdf.pdf.pdfElementer
 import com.itextpdf.kernel.pdf.tagging.StandardRoles
 import com.itextpdf.layout.element.Paragraph
 import com.itextpdf.layout.element.Text
-import no.nav.familie.pdf.pdf.PdfUtils.FontStil
-import no.nav.familie.pdf.pdf.PdfUtils.settFont
+import no.nav.familie.pdf.pdf.FontStil
+import no.nav.familie.pdf.pdf.settFont
 
 fun lagTekstElement(
     tekst: String,
