+++ src/test/kotlin/no/nav/familie/pdf/pdf/InnholdsfortegnelseTest.kt
@@ -1,12 +1,12 @@
 package no.nav.familie.pdf.pdf
 
 import com.itextpdf.layout.element.Div
-import no.nav.familie.pdf.pdf.PdfUtils.håndterRekursivVerdiliste
+import no.nav.familie.pdf.pdf.Innholdsfortegnelse.håndterRekursivVerdiliste
 import no.nav.familie.pdf.pdf.domain.VerdilisteElement
 import org.junit.jupiter.api.Test
 import kotlin.test.assertTrue
 
-class PdfUtilsTest {
+class InnholdsfortegnelseTest {
     @Test
     fun `Dersom label er tom streng skal rekursjonen ikke legge til nytt objekt i seksjonen`() {
         // Arrange
